[jsc#CAP-332] Handle opaque whiteouts as immediately as explicit whiteouts.
The original code defered the deletion until after unpacking completed.
This had the effect of wrongly deleting the just-placed files and
directories, instead of the files and directories of the previous layer.

By deleting immediately the files and directories of the previous layer
are removed, as required. Note that proper operation of this fix relies
on whiteout marker files to come before any new files or directories in
the tar stream.

Note that the existing machinery in the driver was not removed as it is
unclear how much of it is still referenced elsewhere. The fix as-is simply
ensures that this machinery is not triggered.